Data driven documentation on implementations under test
fixes: #45

Adds a GitHub workflow to publish GitHub pages for the site.
The workflow uses the newly added utility to extract the metadata of the validation libraries under test, (extended in this PR to include licence, language and url), and injects this JSON into the markdown files before building and uploading the site.