Split benchmarks into separate build jobs & increase fork count. (#75)
To balance the recent reduction in threads used in benchmarking, the fork count needs to be increased.
This will increase the time it takes each benchmark to complete.
GitHub currently has a 6-hour limit per-job, which would be breached if the fork count was increased.

Therefore, this change splits the benchmark job in two. One job per benchmark.